backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
pandas
osmium